+++ handler/actions/src/main/java/io/knotx/fragments/handler/action/http/HttpActionFactory.java
@@ -15,17 +15,31 @@
  */
 package io.knotx.fragments.handler.action.http;
 
+import static io.vertx.core.http.HttpMethod.DELETE;
+import static io.vertx.core.http.HttpMethod.GET;
+import static io.vertx.core.http.HttpMethod.HEAD;
+import static io.vertx.core.http.HttpMethod.PATCH;
+import static io.vertx.core.http.HttpMethod.POST;
+import static io.vertx.core.http.HttpMethod.PUT;
+
+import com.google.common.collect.ImmutableSet;
 import io.knotx.fragments.handler.action.exception.ActionConfigurationException;
 import io.knotx.fragments.handler.action.http.options.HttpActionOptions;
 import io.knotx.fragments.handler.api.Action;
 import io.knotx.fragments.handler.api.ActionFactory;
 import io.knotx.fragments.handler.api.Cacheable;
 import io.vertx.core.Vertx;
+import io.vertx.core.http.HttpMethod;
 import io.vertx.core.json.JsonObject;
 import io.vertx.reactivex.ext.web.client.WebClient;
+import java.util.Set;
+import java.util.stream.Stream;
 
 @Cacheable
 public class HttpActionFactory implements ActionFactory {
+
+  private static final Set<HttpMethod> SUPPORTED_METHODS = ImmutableSet
+      .of(GET, POST, PATCH, PUT, DELETE, HEAD);
 
   private final WebClientCache webClientCache = new WebClientCache();
 
@@ -36,20 +50,44 @@ public class HttpActionFactory implements ActionFactory {
 
   @Override
   public Action create(String alias, JsonObject config, Vertx vertx, Action doAction) {
+    HttpActionOptions options = new HttpActionOptions(config);
+
+    validateNoDoAction(doAction, alias);
+    validateHttpMethodSupported(options, alias);
+
+    WebClient webClient = webClientCache.getOrCreate(vertx, options.getWebClientOptions());
+    return tryToCreateAction(webClient, options, alias);
+  }
+
+  private void validateNoDoAction(Action doAction, String alias) {
     if (doAction != null) {
       throw new ActionConfigurationException(alias, "Http Action can not wrap another action");
     }
+  }
 
-    HttpActionOptions options = new HttpActionOptions(config);
+  private void validateHttpMethodSupported(HttpActionOptions options, String alias) {
+    HttpMethod method = findHttpMethod(options.getHttpMethod(), alias);
+    if (!SUPPORTED_METHODS.contains(method)) {
+      throw new ActionConfigurationException(alias,
+          String.format("HttpMethod %s configured for HttpAction is not supported",
+              options.getHttpMethod()));
+    }
+  }
 
-    WebClient webClient = webClientCache.getOrCreate(vertx, options.getWebClientOptions());
+  private HttpMethod findHttpMethod(String methodName, String actionAlias) {
+    return Stream.of(HttpMethod.values())
+        .filter(httpMethod -> httpMethod.name().equalsIgnoreCase(methodName))
+        .findFirst()
+        .orElseThrow(() -> new ActionConfigurationException(actionAlias,
+            String.format("HttpMethod %s configured for HttpAction not found in Vert.x library",
+                methodName)));
+  }
 
-    switch (options.getHttpMethod()) {
-      case "GET":
-        return new HttpAction(webClient, options, alias);
-      default:
-        throw new ActionConfigurationException(alias,
-            "HttpMethod configured for HttpAction is not supported");
+  private Action tryToCreateAction(WebClient webClient, HttpActionOptions options, String alias) {
+    try {
+      return new HttpAction(webClient, options, alias);
+    } catch (IllegalArgumentException cause) {
+      throw new ActionConfigurationException(alias, "Creating HttpAction failed", cause);
     }
   }