+++ handler/actions/src/main/java/io/knotx/fragments/handler/action/http/response/EndpointResponse.java
@@ -45,7 +45,6 @@ public class EndpointResponse {
     return endpointResponse;
   }
 
-
   HttpResponseStatus getStatusCode() {
     return statusCode;
   }